pyproject.toml
[project]
name = "llms-overconfidence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
]
